Create env var test (#349)
* Create env var test


* Update core commit


* Fix script


* Try bogus signing key


* Increase wait time


* Increase wait time even more


* Increase wait time to 75


* Try matrix strategy


* Fix test name


* Add check to make sure env path is specified


* Create separate env var tests for v1, v2, and both


* Update .env file in .envrc


---------

+++ scripts/test-proxy-startup-with-env-vars.sh
@@ -2,24 +2,45 @@
 set -e  # Exit on any error
 
 ##### This script is meant to be run in ci #####
-# It tests that the env vars defined in .env.example.holesky are correct.
+# It tests that the env vars defined in the specified environment file are correct.
 # It starts the eigenda-proxy with those env vars, waits 5 seconds, and then kills the proxy.
-# If any deprecated flags are still being used in .env.example.holesky, the script will fail.
+# If any deprecated flags are still being used in the specified environment file, the script will fail.
+
+# Check if an environment file is provided
+if [ $# -eq 0 ]; then
+    echo "Error: No environment file specified"
+    echo "Usage: $0 <environment_file_path>"
+    exit 1
+fi
+
+ENV_FILE=$1
+
+# Check if the environment file exists
+if [ ! -f "$ENV_FILE" ]; then
+    echo "Error: Environment file $ENV_FILE does not exist"
+    echo "Current working directory: $(pwd)"
+    echo "Files in current directory:"
+    ls -la
+
+    exit 1
+fi
+
+echo "Using environment file: $ENV_FILE"
 
 # build the eigenda-proxy binary
 make
 
-# Start the eigenda-proxy with the env vars defined in .env.example.holesky
-set -a; source .env.example.holesky; set +a
+# Start the eigenda-proxy with the env vars defined in the specified environment file
+set -a; source "$ENV_FILE"; set +a
 ./bin/eigenda-proxy &
 PID=$!
 
 # Ensure we kill the process on script exit
 trap "kill $PID" EXIT
 
-# Wait 5 seconds for startup to happen
-echo "sleeping 5 seconds to let the proxy start up"
-sleep 5
+# Wait 75 seconds for startup to happen (actual startup can take up to ~50 seconds, due to SRS parsing)
+echo "sleeping 75 seconds to let the proxy start up"
+sleep 75
 
 echo "Pinging the proxy's health endpoint"
 curl 'http://localhost:3100/health'
